requirements.txt
easyocr==1.4.1
Flask==2.1.1
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
numpy==1.17.4
openalpr==1.1.0
opencv_python==4.5.4.60
Pillow==9.1.0
requests==2.22.0
